+++ environment.yml
@@ -52,5 +52,5 @@ dependencies:
   - pytest >=4.4
   - pytest-cov >=2.6
   - requests-mock >=1.8.0
-  - rasterio
-  - rioxarray
+  - rasterio >=1.2.10
+  - rioxarray >=0.11.1
