+++ test/core/store/fs/impl/test_geotiff.py
@@ -28,8 +28,8 @@ import rioxarray
 import xarray
 import xarray as xr
 
-from xcube.core.store.fs.impl.cog import GeoTIFFMultiLevelDataset
-from xcube.core.store.fs.impl.cog import MultiLevelDatasetGeoTiffFsDataAccessor
+from xcube.core.store.fs.impl.geotiff import GeoTIFFMultiLevelDataset
+from xcube.core.store.fs.impl.geotiff import MultiLevelDatasetGeoTiffFsDataAccessor
 from xcube.core.store.fs.impl.dataset import DatasetGeoTiffFsDataAccessor
 from xcube.util.jsonschema import JsonSchema
 
@@ -39,6 +39,7 @@ class RioXarrayTest(unittest.TestCase):
     This class doesn't test xcube but rather asserts that RioXarray works as
     expected
     """
+
     def test_it_is_possible_get_overviews(self):
         cog_path = os.path.join(os.path.dirname(__file__),
                                 "..", "..", "..", "..", "..",
@@ -92,6 +93,7 @@ class GeoTIFFMultiLevelDatasetTest(unittest.TestCase):
     """
     A class to test wrapping of geotiff file into a multilevel dataset
     """
+
     def test_local_fs(self):
         fs = fsspec.filesystem('file')
         cog_path = os.path.join(os.path.dirname(__file__),
@@ -121,6 +123,7 @@ class MultiLevelDatasetGeoTiffFsDataAccessorTest(unittest.TestCase):
     """
     A class to test cog and GeoTIFF for multilevel dataset opener
     """
+
     def test_read_cog(self):
         fs = fsspec.filesystem('file')
         mldataopener = MultiLevelDatasetGeoTiffFsDataAccessor()
@@ -137,6 +140,8 @@ class MultiLevelDatasetGeoTiffFsDataAccessorTest(unittest.TestCase):
         self.assertEqual("geotiff", mldataopener.get_format_id())
         self.assertIsInstance(mldataopener.get_open_data_params_schema(cog_path)
                               , JsonSchema, "Given object is a instance ")
+        self.assertIsInstance(mldataopener.get_open_data_params_schema(),
+                              JsonSchema)
 
     def test_read_geotiff(self):
         fs = fsspec.filesystem('file')
@@ -154,6 +159,7 @@ class DatasetGeoTiffFsDataAccessorTest(unittest.TestCase):
     A Test class to test opening a GeoTIFF as multilevel dataset or 
     as normal dataset
     """
+
     def test_ml_to_dataset(self):
         fs = fsspec.filesystem('file')
         cog_path = os.path.join(os.path.dirname(__file__),
@@ -172,9 +178,9 @@ class DatasetGeoTiffFsDataAccessorTest(unittest.TestCase):
     def test_dataset(self):
         fs = fsspec.filesystem('file')
         tiff_path = os.path.join(os.path.dirname(__file__),
-                                "..", "..", "..", "..", "..",
-                                "examples", "serve", "demo",
-                                "example-geotiff.tif")
+                                 "..", "..", "..", "..", "..",
+                                 "examples", "serve", "demo",
+                                 "example-geotiff.tif")
         data_accessor = DatasetGeoTiffFsDataAccessor()
         self.assertEqual("geotiff", data_accessor.get_format_id())
         dataset = data_accessor.open_data(data_id=tiff_path, overview_level=0,
@@ -183,4 +189,3 @@ class DatasetGeoTiffFsDataAccessorTest(unittest.TestCase):
         self.assertIsInstance(
             data_accessor.get_open_data_params_schema(tiff_path),
             JsonSchema, "Given object is a instance ")
-
